pyproject.toml
[project]
name = "medical-skin-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Anas-Shaikh-786",email = "anasshaikh78654@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "numpy (<2)",
    "torch (>=2.9.1,<3.0.0)",
    "torchvision (>=0.24.1,<0.25.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "scikit-learn (<1.4)",
    "opencv-python-headless (<4.9)",
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "multipart (>=1.3.0,<2.0.0)",
    "python-multipart (>=0.0.21,<0.0.22)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
